refac: 도메인의 dto에 대한 의존성 제거
