+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,16 +1,13 @@
 package roomescape.controller;
 
 import nextstep.Reservation;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import roomescape.dao.ReservationDAO;
-import roomescape.domain.ReservationRequest;
+import roomescape.dto.ReservationRequest;
 import roomescape.service.ReservationService;
 
 import java.net.URI;
-import java.util.List;
 
 @RestController
 public class ReservationController {
