--- src/main/java/roomescape/domain/ReservationRequest.java
@@ -1,29 +1,0 @@
-package roomescape.domain;
-
-public class ReservationRequest {
-
-    private String date;
-    private String time;
-    private String name;
-
-    public ReservationRequest() {}
-
-    public ReservationRequest(String date, String time, String name) {
-        this.date = date;
-        this.time = time;
-        this.name = name;
-    }
-
-    public String getDate() {
-        return date;
-    }
-
-    public String getTime() {
-        return time;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-}
